test: calibrate dispatch planning functional floor
